refactor(core): extract shared types for function agents
Extract duplicated types from git-func-agent and startchat-func-agent into a shared common module:

- Language enum (Chinese, English)
- AgentError struct with Display and Error traits
- AgentErrorType enum (GitError, AnalysisError, InvalidInput, InternalError)
- AgentResult<T> type alias

Changes:
- Add function_agents/common.rs with shared type definitions
- Re-export shared types from both agent type modules for backward compatibility
- Update imports in ai_service.rs, commit_generator.rs, context_analyzer.rs, work_state_analyzer.rs
- Update function_agents/mod.rs to include common module and re-export shared types

This is a pure refactoring with no functional changes. All existing APIs remain unchanged through re-exports.
