+++ src/crates/core/src/function_agents/startchat-func-agent/ai_service.rs
@@ -1,4 +1,5 @@
 use super::types::*;
+use crate::function_agents::common::{AgentError, AgentResult, Language};
 use crate::infrastructure::ai::AIClient;
 use crate::util::types::Message;
 /**
